+++ src/main/java/com/example/ililbooks/client/dto/GoogleApiResponse.java
@@ -1,0 +1,4 @@
+package com.example.ililbooks.client.dto;
+
+public record GoogleApiResponse() {
+}
